Fix confusion when using type prop to identify children to be rendered
Instead of using the existence of the `type` prop to identify
a child node that needs to be passed to `h`, use a unique symbol
`$$z` to identify them. This avoids accidentally trying to render
children that have already been turned into vnodes elsewhere,
and that vnode has a `type` prop.